feat: point 조회 로직 분리
