+++ src/components/home/Point.tsx
@@ -1,26 +1,19 @@
 import instance from "api/axios";
 import { axiosError } from "api/axiosUtil";
 import { useEffect, useState } from "react";
+import { getPoint } from "hooks/point";
 
 export const Point = () => {
 
-    const [point, setPoint] = useState<number>(0);
-    const viewPoint = point.toString().replace(/\B(?<!\.\d*)(?=(\d{3})+(?!\d))/g, ",");
+    const [point, setPoint] = useState<string>("");
 
-    // GET 포인트 조회
-    const getPoint = async () => {
-        try{
-            const res = await instance.get("/api/main/point");
-            if(res.data.result === "Y"){
-                setPoint(res.data.data[0].point);
-            }
-        }catch(err: any){
-            axiosError(err.message);
-        }
+    const handlePoint = async () => {
+        const usePoint = await getPoint();
+        setPoint(usePoint);
     }
 
     useEffect(() => {
-        getPoint();
+        handlePoint();
     }, []);
 
     return(
@@ -30,7 +23,7 @@ export const Point = () => {
             boxShadow: "20px 20px 60px #bebebe, -20px -20px 60px #ffffff"
         }}>
             <span className="text-[25px]">보유 포인트</span>
-            <span className="text-[25px]"><b className="text-[25px]">{viewPoint}</b> point</span>
+            <span className="text-[25px]"><b className="text-[25px]">{point}</b> point</span>
         </div>
     )
 }
